Fix pull request update showing too many commits with multiple branches
When the base repository contains multiple branches with the same commits as
the base branch, pull requests can show a long list of commits already in the
base branch as having been added.

What this is supposed to do is exclude commits already in the base branch. But
the mechanism to do so assumed a commit only exists in a single branch. Now use
`git rev-list A B --not branchName` instead of filtering commits afterwards.

The logic to detect if there was a force push also was wrong for multiple
branches. If the old commit existed in any branch in the base repository it
would assume there was no force push. Instead check if the old commit is an
ancestor of the new commit.