Add "Reviewed by you" filter for pull requests
This includes pull requests that you approved, requested changes or commented
on. Currently such pull requests are not visible in any of the filters on
/pulls, while they may need further action like merging, or prodding the
author or reviewers.

Especially when working with a large team on a repository it's helpful to get
a full overview of pull requests that may need your attention, without having
to sift through the complete list.