Sort issues and pulls by recently updated in user and organization home
The main purpose of these home pages should be getting an overview of what's
going on or needs attention. Recently updated is a better default than newest
for that purpose, to avoid missing active issues and pulls that were not
created recently.

The default sorting order in repository issues and pulls remains newest.