Improve squash merge commit author and co-author with private emails
When emails addresses are private, squash merges always use @noreply.localhost
for the author of the squash commit. And the author is redundantly added as a
co-author in the commit message.

Also without private mails, the redundant co-author is possible when committing
with a signature that's different than the user full name and primary email.

Now try to find a commit by the same user in the list of commits, and prefer
the signature from that over one constructed from the account settings.