Order patch checker tasks from most to least recently updated
When a change is pushed to the default branch and many pull requests are open
for that branch, conflict checking can take some time to go through all those
pull requests.

Previously it would go from oldest to newest pull request. Now prioritize pull
requests that are likely being actively worked on or prepared for merging.

The change is trivial but also not a full solution:
* There is no ordering between repos, they are handled one after the other.
* If a push happens during patch checking, that PR will end up at the end of
  the queue.

A better solution would be a priority queue for the tasks.