Add goth-provider-blenderid source instead of referencing package
As future.projects.blender.org is going to be removed.

The code is already customized to integrate this provider. Having part of it
live in another repository gets us no closer to using vanilla Gitea. So it's
easier to have it here.

Provider authored by Matti Ranta and Arnd Marijnissen.