Gitealize nicknames from Blender ID
Blender ID allows a wider range of characters in the nickname than Gitea
allows. Before using a Blender ID nickname as a Gitea username, it needs
to be massaged into a valid form.

This was already done in the Blender ID-to-Gitea webhook for username
changes, and this PR introduces it in Gitea itself for the registration
flow.

The implementation follows what the webhook code[1] does, except it's
simpler because it can use built-in Gitea functionality.

This fixes https://projects.blender.org/blender/blender/issues/111937

[1]: https://projects.blender.org/infrastructure/gitea-blenderid-webhook/src/branch/main/gitea_blenderid_webhook/gitea_users.py

Reviewed on: https://github.com/blender/gitea/pull/3